+++ lambda/src/evaluate-script.test.ts
@@ -1,13 +1,3 @@
-/*
- * Copyright 2021 Dgraph Labs, Inc. All rights reserved.
- *
- * Licensed under the Dgraph Community License (the "License"); you
- * may not use this file except in compliance with the License. You
- * may obtain a copy of the License at
- *
- *     https://github.com/dgraph-io/dgraph/blob/master/licenses/DCL.txt
- */
-
 import { evaluateScript } from './evaluate-script';
 import { waitForDgraph, loadSchema, runQuery } from './test-utils'
 import sleep from 'sleep-promise';
